ooo_rms_orderservice/orderservice/requirements.txt
autopep8==2.0.2
#click==8.0.4
Flask==2.0
Flask-JWT-Extended==4.4.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
psycopg2==2.9.5
pycodestyle==2.10.0
PyJWT==2.4.0
Werkzeug==2.0.3